fix(redux-devtools): allow configuring the number of actions to keep in the DevTool (#159)
## Description

Add `maxAge` option to Redux DevTools plugin to configure action history
size. This allows developers to control the maximum number of actions
retained in DevTools; older actions are automatically removed once the
limit is reached (default: 50).

## Related Issue

https://github.com/callstackincubator/rozenite/issues/158

## Context

In our enterprise app, we max out the default 50 actions at boot, so we
appear to be missing actions in DevTools.

## Testing

I use the playground app with the configuration and verify that we
should have more than 50 actions.

---------

Co-authored-by: dprevost <plum117@hotmail.com>
Co-authored-by: Szymon Chmal <szymonchmal@gmail.com>